feat(mcp): implement Model Context Protocol server (ADR-0067)
Implemented the MCP server subcommand `csm mcp serve` to expose CSM as an AI memory provider.
- Added `mcp` feature flag and integrated `rmcp` (official Rust SDK) and `schemars`.
- Implemented 12 memory tools: inject, probe, get, delete, associate, traverse, path, stats, export.
- Supported both Stdio and SSE transports for local and hosted use cases.
- Added comprehensive JSON Schema definitions for tool inputs.
- Provided mdBook documentation and Claude Desktop integration example.
- Fixed CI issues regarding unused imports and unreachable code in the CLI layer.

Note: The implementation targets `rmcp` v0.2.1 and follows its role-based service architecture. Some remaining type mismatches in the MCP model layer may require further alignment with the crate's internal structures.

+++ src/mcp/schema.rs
@@ -1,20 +1,18 @@
+use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
-use rmcp::schemars::JsonSchema;
 
 /// Input for memory_inject tool.
 #[derive(Deserialize, Serialize, JsonSchema, Debug)]
-#[schemars(crate = "rmcp::schemars")]
 pub struct MemoryInjectInput {
     pub id: String,
-    pub vector: Vec<f32>,
+    pub vector: Vec<u8>, // 1280 bytes for 10240 bits
     #[serde(default)]
     pub metadata: HashMap<String, serde_json::Value>,
 }
 
 /// Input for memory_inject_text tool.
 #[derive(Deserialize, Serialize, JsonSchema, Debug)]
-#[schemars(crate = "rmcp::schemars")]
 pub struct MemoryInjectTextInput {
     pub id: String,
     pub text: String,
@@ -24,16 +22,14 @@ pub struct MemoryInjectTextInput {
 
 /// Input for memory_probe tool.
 #[derive(Deserialize, Serialize, JsonSchema, Debug)]
-#[schemars(crate = "rmcp::schemars")]
 pub struct MemoryProbeInput {
-    pub vector: Vec<f32>,
+    pub vector: Vec<u8>,
     #[serde(default = "default_top_k")]
     pub top_k: usize,
 }
 
 /// Input for memory_probe_text tool.
 #[derive(Deserialize, Serialize, JsonSchema, Debug)]
-#[schemars(crate = "rmcp::schemars")]
 pub struct MemoryProbeTextInput {
     pub query: String,
     #[serde(default = "default_top_k")]
@@ -42,7 +38,6 @@ pub struct MemoryProbeTextInput {
 
 /// Input for memory_probe_filtered tool.
 #[derive(Deserialize, Serialize, JsonSchema, Debug)]
-#[schemars(crate = "rmcp::schemars")]
 pub struct MemoryProbeFilteredInput {
     pub query: String,
     #[serde(default = "default_top_k")]
@@ -52,21 +47,18 @@ pub struct MemoryProbeFilteredInput {
 
 /// Input for memory_get tool.
 #[derive(Deserialize, Serialize, JsonSchema, Debug)]
-#[schemars(crate = "rmcp::schemars")]
 pub struct MemoryGetInput {
     pub id: String,
 }
 
 /// Input for memory_delete tool.
 #[derive(Deserialize, Serialize, JsonSchema, Debug)]
-#[schemars(crate = "rmcp::schemars")]
 pub struct MemoryDeleteInput {
     pub id: String,
 }
 
 /// Input for memory_associate tool.
 #[derive(Deserialize, Serialize, JsonSchema, Debug)]
-#[schemars(crate = "rmcp::schemars")]
 pub struct MemoryAssociateInput {
     pub from: String,
     pub to: String,
@@ -76,7 +68,6 @@ pub struct MemoryAssociateInput {
 
 /// Input for memory_traverse tool.
 #[derive(Deserialize, Serialize, JsonSchema, Debug)]
-#[schemars(crate = "rmcp::schemars")]
 pub struct MemoryTraverseInput {
     pub start_id: String,
     #[serde(default = "default_max_depth")]
@@ -87,7 +78,6 @@ pub struct MemoryTraverseInput {
 
 /// Input for memory_shortest_path tool.
 #[derive(Deserialize, Serialize, JsonSchema, Debug)]
-#[schemars(crate = "rmcp::schemars")]
 pub struct MemoryShortestPathInput {
     pub from: String,
     pub to: String,
@@ -95,18 +85,24 @@ pub struct MemoryShortestPathInput {
 
 /// Input for memory_stats tool.
 #[derive(Deserialize, Serialize, JsonSchema, Debug)]
-#[schemars(crate = "rmcp::schemars")]
 pub struct MemoryStatsInput {}
 
 /// Input for memory_export tool.
 #[derive(Deserialize, Serialize, JsonSchema, Debug)]
-#[schemars(crate = "rmcp::schemars")]
 pub struct MemoryExportInput {
     #[serde(default = "default_export_path")]
     pub path: String,
 }
 
-fn default_top_k() -> usize { 10 }
-fn default_strength() -> f32 { 1.0 }
-fn default_max_depth() -> u32 { 3 }
-fn default_export_path() -> String { "export.json".to_string() }
+fn default_top_k() -> usize {
+    10
+}
+fn default_strength() -> f32 {
+    1.0
+}
+fn default_max_depth() -> u32 {
+    3
+}
+fn default_export_path() -> String {
+    "export.json".to_string()
+}
